limit by user, tag crud tweaks

+++ api/src/services/tabs/tabs.ts
@@ -1,3 +1,4 @@
+import { validateWithSync } from '@redwoodjs/api'
 import type {
   QueryResolvers,
   MutationResolvers,
@@ -7,17 +8,24 @@ import type {
 import { db } from 'src/lib/db'
 
 export const tabs: QueryResolvers['tabs'] = () => {
-  return db.tab.findMany()
+  return db.tab.findMany({
+    where: { userId: context.currentUser?.id },
+  })
 }
 
 export const tab: QueryResolvers['tab'] = ({ id }) => {
   return db.tab.findUnique({
-    where: { id },
+    where: { id, userId: context.currentUser?.id },
   })
 }
 
 export const createTab: MutationResolvers['createTab'] = async ({ input }) => {
   const userId = input.userId
+  validateWithSync(() => {
+    if (userId !== context.currentUser?.id) {
+      throw new Error('Unauthorized')
+    }
+  })
   const { tags, ...tabData } = input
   return db.$transaction(async (tx) => {
     const { id: tabId } = await tx.tab.create({
@@ -66,6 +74,11 @@ export const updateTab: MutationResolvers['updateTab'] = async ({
   input,
 }) => {
   const userId = input.userId
+  validateWithSync(() => {
+    if (userId !== context.currentUser?.id) {
+      throw new Error('Unauthorized')
+    }
+  })
   const { tags, ...tabData } = input
   return db.$transaction(async (tx) => {
     await tx.tab.update({ where: { id: tabId }, data: tabData })
@@ -109,6 +122,12 @@ export const updateTab: MutationResolvers['updateTab'] = async ({
 
 export const deleteTab: MutationResolvers['deleteTab'] = ({ id }) => {
   return db.$transaction(async (tx) => {
+    const tab = await tx.tab.findUnique({
+      where: { id, userId: context.currentUser.id },
+    })
+    if (!tab) {
+      return
+    }
     await tx.tabTag.deleteMany({ where: { tabId: id } })
     return tx.tab.delete({ where: { id } })
   })
